Working update from Discores export
Added all new data from discores export, should be able to work again at any time as long as the data is coming from my phone.  Extending past that will take some work, but will be handled later.  When that does get worked on, it should go hand in hand with an exporting function that will export in a form that Discores can import.  Anyway, i disabled the import button so it doesn't get clicked on accident...